Hi,

This pull request introduces optimizations to the `ToTitleCase` method in the `NamingHelperMethods` class to improve performance and reduce memory allocations.

Testing:
-  Passed your unit tests

Key Enhancements:

1. Performance Optimization:
   - Small inputs: 25.4 times faster (4,252.0 ns → 167.3 ns)
   - Medium inputs: 7.3 times faster (105,182.6 ns → 14,419.9 ns)
   - Large inputs: 5.4 times faster (10,983,153.6 ns → 2,048,688.3 ns)
   - Very large inputs: 6.3 times faster (128,103,886.7 ns → 20,264,420.3 ns)

2. Memory Allocation Optimization:
   - Small inputs: 196 times less memory used (9,416 B → 48 B)
   - Medium inputs: 23 times less memory used (45,840 B → 1,992 B)
   - Large inputs: 20 times less memory used (3,926,474 B → 196,977 B)
   - Very large inputs: 18.8 times less memory used (37,083,832 B → 1,968,249 B)

3. Garbage Collection Efficiency:
   - Significant reduction in Gen0, Gen1, and Gen2 collections across all input sizes


![performance_test](https://github.com/user-attachments/assets/bbda7354-2fac-400a-b690-a63fe78d5248)